# Abstract cache interface