--- a/.travis.yml
+++ b/.travis.yml
@@ -1,4 +1,6 @@
-os: osx
+os:
+- linux
+- osx
 language: ruby
 rvm: ruby-2.4.0
 before_script: rake