--- a/bootsnap.gemspec
+++ b/bootsnap.gemspec
@@ -33,4 +33,5 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "ffi-xattr", "~> 0.1.2"
 
   spec.add_runtime_dependency "msgpack", "~> 1.0"
+  spec.add_runtime_dependency "snappy"
 end