--- a/ext/bootsnap/bootsnap.h
+++ b/ext/bootsnap/bootsnap.h
@@ -1,6 +1,0 @@
-#ifndef BOOTSNAP_H
-#define BOOTSNAP_H 1
-
-/* doesn't expose anything */
-
-#endif /* BOOTSNAP_H */