--- a/lib/bootsnap/compile_cache/iseq.rb
+++ b/lib/bootsnap/compile_cache/iseq.rb
@@ -1,64 +1,36 @@
 require 'bootsnap/bootsnap'
-require 'zlib'
+require 'bootsnap/cache/fetch_cache'
 
 module Bootsnap
   module CompileCache
     module ISeq
+      class Cache < FetchCache
+        class << self
+          attr_accessor :compile_option
+        end
+
+        def self.file_key(path)
+          super(path) + [compile_option]
+        end
+      end
+
       class << self
-        attr_accessor :cache_dir
-      end
-
-      def self.input_to_storage(_, path)
-        RubyVM::InstructionSequence.compile_file(path).to_binary
-      rescue SyntaxError
-        raise Uncompilable, 'syntax error'
-      rescue RuntimeError => e
-        if e.message == 'should not compile with coverage'
-          raise Uncompilable, 'coverage is enabled'
-        else
-          raise
-        end
-      end
-
-      def self.storage_to_output(binary)
-        RubyVM::InstructionSequence.load_from_binary(binary)
-      rescue RuntimeError => e
-        if e.message == 'broken binary format'
-          STDERR.puts "[Bootsnap::CompileCache] warning: rejecting broken binary"
-          return nil
-        else
-          raise
-        end
-      end
-
-      def self.input_to_output(_)
-        nil # ruby handles this
+        attr_accessor :cache
       end
 
       module InstructionSequenceMixin
         def load_iseq(path)
-          Bootsnap::CompileCache::Native.fetch(
-            Bootsnap::CompileCache::ISeq.cache_dir,
-            path.to_s,
-            Bootsnap::CompileCache::ISeq
-          )
-        rescue RuntimeError => e
-          if e.message =~ /unmatched platform/
-            puts "unmatched platform for file #{path}"
+          binary = ISeq.cache.fetch(path) do |_, file_path|
+            RubyVM::InstructionSequence.compile_file(file_path).to_binary
           end
-          raise
-        rescue Errno::ERANGE
-          STDERR.puts <<~EOF
-            \x1b[31mError loading ISeq from cache for \x1b[1;34m#{path}\x1b[0;31m!
-            You can likely fix this by running:
-              \x1b[1;32mxattr -c #{path}
-            \x1b[0;31m...but, first, please make sure \x1b[1;34m@burke\x1b[0;31m knows you ran into this bug!
-            He will want to see the results of:
-              \x1b[1;32m/bin/ls -l@ #{path}
-            \x1b[0;31mand:
-              \x1b[1;32mxattr -p user.aotcc.key #{path}\x1b[0m
-          EOF
-          raise
+          RubyVM::InstructionSequence.load_from_binary(binary)
+        rescue RuntimeError => e
+          if e.message == 'broken binary format'
+            STDERR.puts "[Bootsnap::CompileCache] warning: rejecting broken binary"
+            return nil
+          else
+            raise
+          end
         end
 
         def compile_option=(hash)
@@ -68,13 +40,11 @@ module Bootsnap
       end
 
       def self.compile_option_updated
-        option = RubyVM::InstructionSequence.compile_option
-        crc = Zlib.crc32(option.inspect)
-        Bootsnap::CompileCache::Native.compile_option_crc32 = crc
+        Cache.compile_option = RubyVM::InstructionSequence.compile_option.inspect
       end
 
-      def self.install!(cache_dir)
-        Bootsnap::CompileCache::ISeq.cache_dir = cache_dir
+      def self.install!(cache = nil)
+        self.cache = Cache.new(cache)
         Bootsnap::CompileCache::ISeq.compile_option_updated
         class << RubyVM::InstructionSequence
           prepend InstructionSequenceMixin
@@ -83,4 +53,3 @@ module Bootsnap
     end
   end
 end
-