--- a/test/load_path_cache/path_scanner_test.rb
+++ b/test/load_path_cache/path_scanner_test.rb
@@ -1,3 +1,5 @@
+require 'test_helper'
+
 module Bootsnap
   module LoadPathCache
     class PathScannerTest < MiniTest::Test