… README.md
… Rakefile
… ext/bootsnap/bootsnap.c
… ext/bootsnap/extconf.rb