This is a followup to #25.

### Summary

This PR replaces the C-extension compile-cache implementation with an abstract cache interface that allows any cache implementation to be (optionally) provided to `Bootsnap#setup`.

By default, it uses a (slightly modified) existing `load_path_cache/store` implementation, which persists a serialized/compressed Ruby hash to a single file when the process exits.

### Benchmarks

I ran some measurements of `discourse` load-times on my local Ubuntu system comparing this PR against the current `master` branch, and there was little performance difference, while the use of `snappy` to persist the cache-storage file results in substantial disk-space savings:

Average of 5 test-runs:
`extract_cache` branch: **2.893ms** save; **1.708ms** load; **13M** disk space
`master` branch: **2.957ms** save; **1.716ms** load; **45M** disk space

(Detailed output in [gist](https://gist.github.com/wjordan/91f577f5af06c5e05a9c083dc3ccd5a3).)