@wjordan 2017-06-06
I started some work on supporting the native c-extension fetch-cache (as opposed to removing it entirely) in another branch: [`extract_cache_c_ext`](https://github.com/wjordan/bootsnap/compare/extract_cache...wjordan:extract_cache_c_ext?expand=1) - I had difficulty getting it to work reliably without segfaults, possibly related to the slightly modified interface - the iseq compile cache was segfaulting on very-large sources ([`certificate_generator.rb`](https://github.com/discourse/discourse/blob/master/plugins/discourse-narrative-bot/lib/discourse_narrative_bot/certificate_generator.rb) in `discourse`), and it was having all sorts of 'invalid binary' issues unless you run msgpack before storing the iseq binary.