@wjordan 2017-07-05
@burke please take a look and let me know if you're willing to even consider merging a PR along these lines. Happy to engage in further discussion or follow-up work if you have specific code comments or concerns with the general approach.