@burke 2017-07-12
Just getting back from a vacation. I'll take a look at this in the next little bit, lots to catch up on. I'm not immediately opposed to the idea.